+++ fabric/src/main/java/tocraft/craftedcore/registration/fabric/SynchronizedReloadListenerRegistryImpl.java
@@ -26,10 +26,17 @@ public class SynchronizedReloadListenerRegistryImpl {
                 return reloadListener.getName();
             }
 
+            //#if MC>=1212
             @Override
-            public @NotNull CompletableFuture<Void> reload(PreparationBarrier preparationBarrier, ResourceManager resourceManager, ProfilerFiller preparationsProfiler, ProfilerFiller reloadProfiler, Executor backgroundExecutor, Executor gameExecutor) {
-                return reloadListener.reload(preparationBarrier, resourceManager, preparationsProfiler, reloadProfiler, backgroundExecutor, gameExecutor);
+            public @NotNull CompletableFuture<Void> reload(PreparationBarrier preparationBarrier, ResourceManager resourceManager, Executor executor, Executor executor2) {
+                return reloadListener.reload(preparationBarrier, resourceManager, executor, executor2);
             }
+            //#else
+            //$$ @Override
+            //$$ public @NotNull CompletableFuture<Void> reload(PreparationBarrier preparationBarrier, ResourceManager resourceManager, ProfilerFiller preparationsProfiler, ProfilerFiller reloadProfiler, Executor backgroundExecutor, Executor gameExecutor) {
+            //$$     return reloadListener.reload(preparationBarrier, resourceManager, preparationsProfiler, reloadProfiler, backgroundExecutor, gameExecutor);
+            //$$ }
+            //#endif
         });
     }
 }